Finalizacion de servicios necesarios y controladores en trabajo

+++ src/main/java/com/soquetes/Loquetes/entidades/Detalle_compra.java
@@ -26,6 +26,6 @@ public class Detalle_compra {
 
     private Integer cantidad;
 
-    private Double precio;
+    private Integer precio;
 
 }
